Move zendesk-relevant models to zendesk namespace
