+++ lib/zendesk/snippet_collection.rb
@@ -1,0 +1,12 @@
+module Zendesk
+  class SnippetCollection
+    def initialize(snippets)
+      @snippets = snippets
+    end
+
+    def to_s
+      applicable_snippets = @snippets.select(&:applies?)
+      applicable_snippets.collect(&:to_s).join("\n\n")
+    end
+  end
+end
